feat: move sprocket rating calc to its own service
